test: 도메인별로 흩어져있는 통합테스트를 하나로 합친다
