+++ src/test/java/net/teumteum/integration/MeetingIntegrationTest.java
@@ -1,5 +1,6 @@
-package net.teumteum.meeting.integration;
+package net.teumteum.integration;
 
+import java.util.List;
 import net.teumteum.core.error.ErrorResponse;
 import net.teumteum.meeting.domain.ResultCursor;
 import net.teumteum.meeting.domain.response.MeetingResponse;
@@ -9,10 +10,9 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.core.ParameterizedTypeReference;
 
-import java.util.List;
-
 @DisplayName("유저 통합테스트의")
 class MeetingIntegrationTest extends IntegrationTest {
+
     private static final String VALID_TOKEN = "VALID_TOKEN";
     private static final String INVALID_TOKEN = "IN_VALID_TOKEN";
     public static final Long FIRST_REQUEST_CURSOR_ID = 0L;
@@ -20,6 +20,7 @@ class MeetingIntegrationTest extends IntegrationTest {
     @Nested
     @DisplayName("단일 미팅 조회 API는")
     class Find_meeting_api {
+
         @Test
         @DisplayName("존재하는 모임의 id가 주어지면, 모임 정보를 응답한다.")
         void Return_meeting_info_if_exist_meeting_id_received() {
@@ -30,11 +31,11 @@ class MeetingIntegrationTest extends IntegrationTest {
             var result = api.getMeetingById(VALID_TOKEN, meeting.getId());
             // then
             Assertions.assertThat(
-                            result.expectStatus().isOk()
-                                    .expectBody(MeetingResponse.class)
-                                    .returnResult().getResponseBody())
-                    .usingRecursiveComparison()
-                    .isEqualTo(expected);
+                    result.expectStatus().isOk()
+                        .expectBody(MeetingResponse.class)
+                        .returnResult().getResponseBody())
+                .usingRecursiveComparison()
+                .isEqualTo(expected);
         }
 
         @Test
@@ -46,13 +47,14 @@ class MeetingIntegrationTest extends IntegrationTest {
             var result = api.getMeetingById(VALID_TOKEN, notExistMeetingId);
             // then
             result.expectStatus().isBadRequest()
-                    .expectBody(ErrorResponse.class);
+                .expectBody(ErrorResponse.class);
         }
     }
 
     @Nested
     @DisplayName("미팅 목록 조회 API는")
     class Find_meeting_list_api {
+
         @Test
         @DisplayName("첫 페이지 요청에 주어진 size 만큼 미팅 목록을 최신순으로 응답한다.")
         void Return_meeting_list_if_0_cursor_id_and_size_received() {
@@ -61,21 +63,21 @@ class MeetingIntegrationTest extends IntegrationTest {
             var meetings = repository.saveAndGetOpenMeetings(existSize);
 
             List<MeetingResponse> expectedData = meetings.reversed().stream()
-                    .map(MeetingResponse::of)
-                    .toList();
+                .map(MeetingResponse::of)
+                .toList();
             var expected = ResultCursor.create(expectedData, expectedData.getLast().id(), existSize);
             var requestSize = 5;
             // when
             var result = api.getOpenMeetings(VALID_TOKEN, FIRST_REQUEST_CURSOR_ID, requestSize);
             // then
             Assertions.assertThat(
-                            result.expectStatus().isOk()
-                                    .expectBody(new ParameterizedTypeReference<ResultCursor<MeetingResponse>>() {
-                                    })
-                                    .returnResult().getResponseBody())
-                    .usingRecursiveComparison()
-                    .ignoringFields("hasNext", "cursorId")
-                    .isEqualTo(expected);
+                    result.expectStatus().isOk()
+                        .expectBody(new ParameterizedTypeReference<ResultCursor<MeetingResponse>>() {
+                        })
+                        .returnResult().getResponseBody())
+                .usingRecursiveComparison()
+                .ignoringFields("hasNext", "cursorId")
+                .isEqualTo(expected);
         }
 
         @Test
@@ -87,21 +89,22 @@ class MeetingIntegrationTest extends IntegrationTest {
 
             var requestSize = 5;
             var expectedData = meetings.reversed().subList(0, requestSize).stream()
-                    .map(MeetingResponse::of)
-                    .toList();
+                .map(MeetingResponse::of)
+                .toList();
             var cursorId = expectedData.getFirst().id();
-            ResultCursor<MeetingResponse> expected = ResultCursor.create(expectedData, expectedData.getLast().id(), requestSize);
+            ResultCursor<MeetingResponse> expected = ResultCursor.create(expectedData, expectedData.getLast().id(),
+                requestSize);
             // when
             var result = api.getOpenMeetings(VALID_TOKEN, cursorId, requestSize);
             // then
             Assertions.assertThat(
-                            result.expectStatus().isOk()
-                                    .expectBody(new ParameterizedTypeReference<ResultCursor<MeetingResponse>>() {
-                                    })
-                                    .returnResult().getResponseBody())
-                    .usingRecursiveComparison()
-                    .ignoringFields("hasNext", "cursorId")
-                    .isEqualTo(expected);
+                    result.expectStatus().isOk()
+                        .expectBody(new ParameterizedTypeReference<ResultCursor<MeetingResponse>>() {
+                        })
+                        .returnResult().getResponseBody())
+                .usingRecursiveComparison()
+                .ignoringFields("hasNext", "cursorId")
+                .isEqualTo(expected);
         }
     }
 }
